Changes to get things to (mostly) build on FreeBSD.
I have disabled building the documentation. It isn't really important to what I'm trying to do and it was throwing troff errors.  I have not tried to preserve backwards compatibility, so it is likely things will break if you want to find a 26 year old UNIX version on which it used to work and try building it there.